+++ src/Task/String/ToUpperCaseTask.php
@@ -3,20 +3,9 @@ namespace Strivex\Phing\Task\String;
 
 use Phing\Task;
 
-class ToUpperCaseTask extends Task {
-    
-    private $input;
-    private $propertyName;
-    
-    public function setInput($value) {
-        $this->input = $value;
-    }
-    
-    public function setPropertyName($value) {
-        $this->propertyName = $value;
-    }
+class ToUpperCaseTask extends AbstractStringCaseTask {
     
     public function main() {
-        $this->getProject()->setProperty($this->propertyName, strtoupper($this->input));
+        $this->getProject()->setProperty($this->propertyName, $this->toUpperCase($this->input));
     }
 }