+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# Version 1.0.0
+- Using Strivex\Commons
+- Added custom Phing tasks
+  - Strivex\Phing\Task\String\
+    - ToPascalCaseTask
+    - ToCamelCaseTask
+    - ToSnakeCaseTask
+    - ToKebabCaseTask
+    - StringCaseTask
+- Updated README.md
+- Updated CHANGELOG.md
+- Added docs/task/String.md
+
 # Version 0.1.0
 - First release.
 - Added custom Phing tasks